Fix gfx1201 queue topology and SDMA ring wraparound
Map seven persistent queues across the two four-queue compute pipes and preserve monotonic 64-bit SDMA write pointers. Gate public queue capabilities on verified driver 187 support.

Validate all queue slots, multi-producer wraparound, two-process ownership and continued dispatch after peer exit, plus concurrent shader/CP/CPU-HSA signals. Record hardware results and remaining HRX integration requirements.

+++ dext/MacAMDGPU.cpp
@@ -2401,7 +2401,7 @@ MacAMDGPUUserClient::ExternalMethod(uint64_t selector,
         auto &bringup = driver->ivars->bringup;
         const auto previousPSPFence = bringup.psp.fenceCounter;
         const auto previousCPWptr = bringup.cp.wptr;
-        uint32_t previousSDMAWptr[amdgpu::kSDMAInstanceCount] = {};
+        uint64_t previousSDMAWptr[amdgpu::kSDMAInstanceCount] = {};
         for (uint32_t i = 0; i < amdgpu::kSDMAInstanceCount; ++i)
             previousSDMAWptr[i] = bringup.sdma.instance[i].wptr;
         kern_return_t ret = amdgpu::bringup_to(bringup, stage);
